commit 15: move describeAction + state types into src/logic/
Extract ClueAction / ClueState / DraftSuggestion / UiMode types out
of state.tsx into src/logic/ClueState.ts — they describe the
reducer's data shape, not its React plumbing, and decoupling them
unblocks pure-logic consumers that were previously stuck importing
across the ui/logic boundary.

state.tsx now re-exports all four symbols for back-compat so
external importers (`SuggestionLogPanel` reads DraftSuggestion from
state) don't break.

Move src/ui/describeAction.ts -> src/logic/describeAction.ts:
- Imports swap from "./state" / "../logic/GameSetup" to
  "./ClueState" / "./GameSetup" now that ClueAction + ClueState live
  in the logic layer.
- Toolbar.tsx updates its import path ("../describeAction" ->
  "../../logic/describeAction"). Only UI consumer.

Also drops CardCategory and Card imports from state.tsx that had
been pulling through describeAction's now-deleted re-export chain;
Card gets re-added because the reducer body still uses it in a
session-replace branch.

Verify: pnpm typecheck / test / lint / knip clean. Browser: Toolbar
Undo/Redo buttons render with their static aria-label and the
"Undo (⌘Z / Ctrl+Z)" shortcut hint; the dynamic describeAction
tooltip fires on hover — same describer code, same i18n keys, same
copy, just living in the logic layer now.
